[BOOKINGSG-9134] rename the dropdown list v2 and update file import
