+++ src/input-group/input-group-list-addon.style.tsx
@@ -1,6 +1,6 @@
 import styled, { css } from "styled-components";
 import { Input } from "../input";
-import { ExpandableElement } from "../shared/dropdown-list-v2";
+import { ExpandableElement } from "../shared/dropdown-list";
 import { InputBox } from "../shared/input-wrapper";
 import { Colour, Spacing } from "../theme";
 import { InputGroupAddonPosition } from "./types";
