+++ src/input-group/input-group-list-addon.tsx
@@ -1,7 +1,7 @@
 import { OpenChangeReason } from "@floating-ui/react";
 import React, { useEffect, useRef, useState } from "react";
 import { VisuallyHidden, concatIds } from "../shared/accessibility";
-import { DropdownList, DropdownListState } from "../shared/dropdown-list-v2";
+import { DropdownList, DropdownListState } from "../shared/dropdown-list";
 import { ElementWithDropdown } from "../shared/dropdown-wrapper";
 import {
     LabelContainer,
