+++ src/input-multi-select/input-multi-select.tsx
@@ -5,7 +5,7 @@ import {
     DropdownList,
     DropdownListState,
     ExpandableElement,
-} from "../shared/dropdown-list-v2";
+} from "../shared/dropdown-list";
 import { ElementWithDropdown } from "../shared/dropdown-wrapper";
 import {
     LabelContainer,
