+++ src/input-nested-multi-select/input-nested-multi-select.tsx
@@ -8,7 +8,7 @@ import {
     NestedDropdownListItemProps,
     NestedDropdownListLocalItem,
     buildKeyPathToSet,
-} from "../shared/dropdown-list-v2";
+} from "../shared/dropdown-list";
 import { ElementWithDropdown } from "../shared/dropdown-wrapper";
 import {
     LabelContainer,
