+++ src/input-nested-select/types.ts
@@ -8,7 +8,7 @@ import {
     DropdownSearchProps,
     DropdownVariantType,
     ExpandMode,
-} from "../shared/dropdown-list-v2/types";
+} from "../shared/dropdown-list/types";
 import { DropdownAlignmentType } from "../shared/dropdown-wrapper/types";
 
 // =============================================================================
