+++ src/input-range-select/input-range-select.tsx
@@ -3,7 +3,7 @@ import {
     DropdownList,
     DropdownListApi,
     DropdownListState,
-} from "../shared/dropdown-list-v2";
+} from "../shared/dropdown-list";
 import { ElementWithDropdown } from "../shared/dropdown-wrapper";
 import {
     PlaceholderLabel,
