+++ src/input-range-select/types.ts
@@ -6,7 +6,7 @@ import {
     DropdownDisplayProps,
     DropdownSearchProps,
     ItemsLoadStateType,
-} from "../shared/dropdown-list-v2/types";
+} from "../shared/dropdown-list/types";
 import { RefObject } from "react";
 import { DropdownAlignmentType } from "src/shared/dropdown-wrapper";
