+++ src/input-select/input-select.tsx
@@ -4,7 +4,7 @@ import {
     DropdownList,
     DropdownListState,
     ExpandableElement,
-} from "../shared/dropdown-list-v2";
+} from "../shared/dropdown-list";
 import { ElementWithDropdown } from "../shared/dropdown-wrapper";
 import {
     LabelContainer,
