+++ src/input-select/types.ts
@@ -6,7 +6,7 @@ import {
     DropdownVariantType,
     ItemsLoadStateType,
     TruncateType,
-} from "../shared/dropdown-list-v2/types";
+} from "../shared/dropdown-list/types";
 import { DropdownAlignmentType } from "../shared/dropdown-wrapper";
 
 // =============================================================================
