+++ src/shared/dropdown-list/expandable-element.styles.tsx
